Changed the return type of getHight and getWidth in ICollidable to double

+++ project/HowDoIFlyThisThing/src/main/java/com/whathappensingandalf/howdoiflythisthing/Asteroid.java
@@ -26,11 +26,11 @@ public class Asteroid implements ICollidable{
 		throw new UnsupportedOperationException("Not supported yet."); //To change body of generated methods, choose Tools | Templates.
 	}
 
-	public int getHeight() {
-		throw new UnsupportedOperationException("Not supported yet."); //To change body of generated methods, choose Tools | Templates.
+	public double getHeight() {
+				throw new UnsupportedOperationException("Not supported yet."); //To change body of generated methods, choose Tools | Templates.
 	}
 
-	public int getWidth() {
+	public double getWidth() {
 		throw new UnsupportedOperationException("Not supported yet."); //To change body of generated methods, choose Tools | Templates.
 	}
