+++ project/HowDoIFlyThisThing/src/main/java/com/whathappensingandalf/howdoiflythisthing/ICollidable.java
@@ -18,11 +18,11 @@ public interface ICollidable extends IListable{
 	/**
 	 * @return the height of this component
 	 */
-	public int getHeight();
+	public double getHeight();
 	/**
 	 * @return the width of this component
 	 */
-	public int getWidth();
+	public double  getWidth();
 	/**
 	 * @return the position of this component
 	 */
